kchive/requirements.txt
asgiref==3.5.2
astroid==2.5.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.9.3
bs4==0.0.1
bwsi-grader==1.11.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.1.0
colorama==0.4.4
commonmark==0.9.1
configparser==5.2.0
crayons==0.4.0
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-allauth==0.50.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
django-rest-knox==4.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
firebase-admin==5.0.1
future==0.18.2
google-api-core==1.31.0
google-api-python-client==2.14.0
google-auth==1.33.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.2.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
googletrans==4.0.0rc1
graphviz==0.20
grpcio==1.39.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpie==3.2.1
httplib2==0.19.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==5.8.0
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
matplotlib==3.4.1
mccabe==0.6.1
msgpack==1.0.2
multidict==6.0.2
mysql-connector-python==8.0.27
numpy==1.20.2
oauthlib==3.2.0
outcome==1.1.0
packaging==21.0
pandas==1.2.3
patsy==0.5.1
Pillow==8.2.0
pipi==1.0.1
proto-plus==1.19.0
protobuf==3.17.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.12.0
PyJWT==1.7.1
pylint==2.7.4
pyOpenSSL==21.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-firebase==1.2
python-twitter==3.5
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==12.4.4
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.3
seaborn==0.11.1
selenium==4.1.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
statsmodels==0.12.2
threadpoolctl==2.1.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.10.0
uritemplate==3.0.1
urllib3==1.26.4
webdriver-manager==3.5.2
whitenoise==5.3.0
wrapt==1.12.1
wsproto==1.0.0
xlrd==2.0.1
